Basic tests + interfaces added for fork() and map() composite functions.
