+++ Composite/Composite.Cs.Tests/Enumerable/AsEnumerableTests.cs
@@ -8,27 +8,8 @@ using Composite;
 
 namespace Composite.Cs.Tests
 {
-    public class Tests
+    public class AsEnumerableTests
     {
-        [Fact]
-        public void ToForestTest(){
-            var obj = C.Composite (new [] {
-                C.Value(new Simple { Number = 2, }),
-                C.Composite(new [] {
-                    C.Value(new Simple { Number = 1, }),
-                    C.Value(new Simple { Number = 6, }),
-                }),
-                C.Value(new Simple { Number = 2, }),
-            }.AsLimited(2));
-
-            obj.ToForest().Take(2).ToArray();
-
-            Assert.Throws<InvalidOperationException>(() =>
-            {
-                obj.ToForest().Take(3).ToArray();
-            });
-        }
-
         [Fact]
         public void AsEnumerableTest()
         {
