+++ Composite/Composite/Composite.fsproj
@@ -14,7 +14,6 @@
     <Compile Include="Seq.fs" />
     <Compile Include="Comp.fs" />
     <Compile Include="CheckTransformRule.fs"/>
-    <Compile Include="CheckUnfoldRule.fs"/>
     <Compile Include="CompositeExtensions.fs" />
     <Compile Include="EnumerableExtensions.fs" />
     <Compile Include="C.fs" />
